pyproject.toml
[project]
name = "dftlabs-backend"
version = "0.1.0"
description = "DeepFly Tech Labs — Production API"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.22.1",
    "anthropic>=0.84.0",
    "apscheduler>=3.11.2",
    "arxiv>=2.4.0",
    "asyncpg>=0.31.0",
    "beautifulsoup4>=4.14.3",
    "fastapi>=0.135.1",
    "greenlet>=3.3.2",
    "httpx>=0.28.1",
    "openai>=2.24.0",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2>=2.9.11",
    "pydantic-settings>=2.13.1",
    "pydantic[email]>=2.12.5",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.22",
    "resend>=2.23.0",
    "sqlmodel>=0.0.37",
    "uvicorn[standard]>=0.41.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.15.4",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501", "B008"]

[tool.mypy]
python_version = "3.12"
strict = false
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
# removed: env_files = [".env.test"]

[tool.coverage.report]
exclude_lines = ["pragma: no cover", "if TYPE_CHECKING:"]


